Added axialSrc.m for bottle-brush simulations
axialSrc.m parses an outifile for the MassColors section of the header. Converts the red bead, labeled as 1, to 4s and all other colors/integers to 0. The converted string of integers is written to a text file. The resulting src file can be used to generate fluorescent images using the brownianToFluorosim Python codes.